requirements.txt
ccxt==4.4.3
PyYAML==6.0.2
requests==2.32.3
